requirements.txt
numpy
torch
opencv-python
pillow
random2
scikit-learn
torchvision
tqdm
glob
thinker